requirements.txt
mysql-connector-python==9.2.0
psycopg[binary]==3.2.6
